Replace socket.io with rocketchat:streamer
Nothing is wrong with socket.io itself.. But for the upcoming milestone,
project needs to be deployed to Galaxy, and Galaxy has a problem with
socket.io. Galaxy only allows one external port to be used by the
application. socket.io needs to sit on a port different from the
default port that maps 80 to itself. There are many examples online that
demonstrates having both servers on the same port with Express.. But we couldn't
figure out a way to replicate that on Meteor.