Add ping system to Game class and game page
Ping system refers to a system in which the client sends empty signals
to the server periodically to indicate that it is still active. If the
server doesn't receive these signals from a client, then we can safely
remove that client from the game. This is more efficient than listening
for a browser refresh, back button press, etc. to determine whether a
client has disconnected or not.