Change game start logic
Before applying this change, there is a readyCount member variable that
keeps track of the number of ready signals fired in a lobby. Upon
readySignal method call by the client, we increment readyCount and check
if readyCount equals the number of players in the room. If true, we
start the game.

There are two known problems to this approach:
1. A player cannot revert their ready state.
2. One player can fire multiple ready signals to start the game in a
   lobby with multiple players.

After this change, readyCount will be eliminated. Instead, when ready signal is
fired (whose method is now called toggleReady), the ready switch property of
that player will toggle. We then check if this property is true for all
of the players and start the game if true.